// js/stat-evaluation-player/src/generated/TerritorialPressureEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TerritorialPressureEndReason } from "./TerritorialPressureEndReason.ts";

export type TerritorialPressureEvent = { start_time: number, start_frame: number, end_time: number, end_frame: number, team_is_team_0: boolean, duration: number, offensive_half_time: number, offensive_third_time: number, end_reason: TerritorialPressureEndReason, };