// js/stat-evaluation-player/src/generated/TerritorialPressureTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TerritorialPressureTeamStats = { tracked_time: number, session_count: number, opponent_session_count: number, session_time: number, opponent_session_time: number, offensive_half_time: number, offensive_third_time: number, longest_session_time: number, opponent_longest_session_time: number, average_session_time: number, };